Extract spawn robots form otto_fleet_nav
