+++ ros2_ws/src/roscon2023_demo/bash/spawn.sh
@@ -9,8 +9,7 @@ export ConfigDir=${SCRIPTPATH}/../../roscon2023_demo/config/
 
 echo Starting MoveIt stacks and deliberation locks
 
-screen -S MOVEIT_LOCKS -dm ros2 launch roscon2023_demo ROSCon2023Demo.launch.py ROS2Con2023Config:=${ConfigDir}/ROSCon2023Config_900.yaml > /tmp/moveit.log
-
+screen -S MOVEIT_LOCKS -dm ros2 launch roscon2023_demo ROSCon2023Demo.launch.py ROS2Con2023Config:=${ConfigDir}/ROSCon2023Config_900.yaml start_amr_navigation:=False spawn_amrs:=False
 
 ros2 service call /spawn_entity gazebo_msgs/srv/SpawnEntity "{name: 'otto600', robot_namespace: 'otto_1', initial_pose: {position: {x: -32.0 , y: 2, z: 0.1}}}"
 ros2 service call /spawn_entity gazebo_msgs/srv/SpawnEntity "{name: 'otto600', robot_namespace: 'otto_2', initial_pose: {position: {x: -30.0 , y: 2, z: 0.1}}}"
@@ -42,5 +41,4 @@ ros2 service call /spawn_entity gazebo_msgs/srv/SpawnEntity "{name: 'otto600', r
 ros2 service call /spawn_entity gazebo_msgs/srv/SpawnEntity "{name: 'otto600', robot_namespace: 'otto_36', initial_pose: {position: {x: -30.0 , y: 87, z: 0.1}}}"
 ros2 service call /spawn_entity gazebo_msgs/srv/SpawnEntity "{name: 'otto600', robot_namespace: 'otto_37', initial_pose: {position: {x: -28.0 , y: 87, z: 0.1}}}"
 ros2 service call /spawn_entity gazebo_msgs/srv/SpawnEntity "{name: 'otto600', robot_namespace: 'otto_38', initial_pose: {position: {x: -26.0 , y: 87, z: 0.1}}}"
-
