+++ ros2_ws/src/roscon2023_demo/config/ROSCon2023Config_900.yaml
@@ -174,15 +174,15 @@ fleet:
 
 arms:
   - namespace: ur1
-    num_of_boxes: 6
+    num_of_boxes: 18
     launch_rviz: false
 
   - namespace: ur2
-    num_of_boxes: 6
+    num_of_boxes: 18
     launch_rviz: false
 
   - namespace: ur3
-    num_of_boxes: 6
+    num_of_boxes: 18
     launch_rviz: false
 
   - namespace: ur4
