--- ros2_ws/src/roscon2023_demo/config/ROSCon2023Config_900_group1.yaml
@@ -1,99 +1,0 @@
-fleet:
-# Group 1
-  - robot_name: otto600
-    robot_namespace: otto_1
-    lane: Line1
-    tasks_config_file: config/otto600_tasks_12.yaml
-    nav2_param_file: otto600Params.yaml
-    position: 
-      x: -32.0
-      y: 2.0
-      z: 0.1
-
-  - robot_name: otto600
-    robot_namespace: otto_2
-    lane: Line1
-    tasks_config_file: config/otto600_tasks_12.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -30.0
-      y: 2.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_3
-    lane: Line2
-    tasks_config_file: config/otto600_tasks_12.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -28.0
-      y: 2.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_4
-    lane: Line2
-    tasks_config_file: config/otto600_tasks_12.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -26.0
-      y: 2.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_5
-    lane: Line3
-    tasks_config_file: config/otto600_tasks_34.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -32.0
-      y: 27.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_6
-    lane: Line3
-    tasks_config_file: config/otto600_tasks_34.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -30.0
-      y: 27.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_7
-    lane: Line4
-    tasks_config_file: config/otto600_tasks_34.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -28.0
-      y: 27.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_8
-    lane: Line4
-    tasks_config_file: config/otto600_tasks_34.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -26.0
-      y: 27.0
-      z: 0.2
-
-arms:
-# Group 1
-  - namespace: ur1
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur2
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur3
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur4
-    num_of_boxes: 18
-    launch_rviz: false
