--- ros2_ws/src/roscon2023_demo/config/ROSCon2023Config_900_group2.yaml
@@ -1,100 +1,0 @@
-fleet:
-  
-  - robot_name: otto600
-    robot_namespace: otto_21
-    lane: Line21
-    tasks_config_file: config/otto600_tasks_12_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position: 
-      x: -32.0
-      y: 32.0
-      z: 0.1
-
-  - robot_name: otto600
-    robot_namespace: otto_22
-    lane: Line21
-    tasks_config_file: config/otto600_tasks_12_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -30.0
-      y: 32.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_23
-    lane: Line22
-    tasks_config_file: config/otto600_tasks_12_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -28.0
-      y: 32.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_24
-    lane: Line22
-    tasks_config_file: config/otto600_tasks_12_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -26.0
-      y: 32.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_25
-    lane: Line23
-    tasks_config_file: config/otto600_tasks_34_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -32.0
-      y: 57.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_26
-    lane: Line23
-    tasks_config_file: config/otto600_tasks_34_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -30.0
-      y: 57.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_27
-    lane: Line24
-    tasks_config_file: config/otto600_tasks_34_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -28.0
-      y: 57.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_28
-    lane: Line24
-    tasks_config_file: config/otto600_tasks_34_lane2.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -26.0
-      y: 57.0
-      z: 0.2
-
-arms:
-# Group 2
-
-  - namespace: ur5
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur6
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur7
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur8
-    num_of_boxes: 18
-    launch_rviz: false
