--- ros2_ws/src/roscon2023_demo/config/ROSCon2023Config_900_group3.yaml
@@ -1,100 +1,0 @@
-fleet:
-  - robot_name: otto600
-    robot_namespace: otto_31
-    lane: Line31
-    tasks_config_file: config/otto600_tasks_12_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position: 
-      x: -32.0
-      y: 62.0
-      z: 0.1
-
-  - robot_name: otto600
-    robot_namespace: otto_32
-    lane: Line31
-    tasks_config_file: config/otto600_tasks_12_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -30.0
-      y: 62.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_33
-    lane: Line32
-    tasks_config_file: config/otto600_tasks_12_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -28.0
-      y: 62.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_34
-    lane: Line32
-    tasks_config_file: config/otto600_tasks_12_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -26.0
-      y: 62.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_35
-    lane: Line33
-    tasks_config_file: config/otto600_tasks_34_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -32.0
-      y: 87.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_36
-    lane: Line33
-    tasks_config_file: config/otto600_tasks_34_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -30.0
-      y: 87.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_37
-    lane: Line34
-    tasks_config_file: config/otto600_tasks_34_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -28.0
-      y: 87.0
-      z: 0.2
-
-  - robot_name: otto600
-    robot_namespace: otto_38
-    lane: Line34
-    tasks_config_file: config/otto600_tasks_34_lane3.yaml
-    nav2_param_file: otto600Params.yaml
-    position:
-      x: -26.0
-      y: 87.0
-      z: 0.2
-      
-
-arms:
-# Group 3
-
-  - namespace: ur9
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur10
-    num_of_boxes: 18
-    launch_rviz: false   
-
-  - namespace: ur11
-    num_of_boxes: 18
-    launch_rviz: false
-
-  - namespace: ur12
-    num_of_boxes: 18
-    launch_rviz: false
